+++ src/HotChocolate/Core/src/Types.Scalars/HslType.cs
@@ -1,0 +1,114 @@
+using System.Text.RegularExpressions;
+using HotChocolate.Language;
+
+namespace HotChocolate.Types.Scalars
+{
+    /// <summary>
+    /// The `HSL` scalar type represents a valid a CSS HSL color as defined
+    /// here https://developer.mozilla.org/en-US/docs/Web/CSS/color_value#hsl_colors.
+    /// </summary>
+    public class HslType : StringType
+    {
+        private static readonly string _validationPattern =
+            ScalarResources.HslType_ValidationPattern;
+
+        protected static readonly Regex ValidationRegex =
+            new(_validationPattern, RegexOptions.Compiled | RegexOptions.IgnoreCase);
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="HslType"/> class.
+        /// </summary>
+        public HslType()
+            : this(
+                WellKnownScalarTypes.Hsl,
+                ScalarResources.HslType_Description)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="HslType"/> class.
+        /// </summary>
+        public HslType(
+            NameString name,
+            string? description = null,
+            BindingBehavior bind = BindingBehavior.Explicit)
+            : base(name, description, bind)
+        {
+            Description = description;
+        }
+
+        /// <inheritdoc />
+        protected override bool IsInstanceOfType(string runtimeValue)
+        {
+            return ValidationRegex.IsMatch(runtimeValue);
+        }
+
+        /// <inheritdoc />
+        protected override bool IsInstanceOfType(StringValueNode valueSyntax)
+        {
+            return ValidationRegex.IsMatch(valueSyntax.Value);
+        }
+
+        /// <inheritdoc />
+        protected override string ParseLiteral(StringValueNode valueSyntax)
+        {
+            if (!ValidationRegex.IsMatch(valueSyntax.Value))
+            {
+                throw ThrowHelper.HslType_ParseLiteral_IsInvalid(this);
+            }
+
+            return base.ParseLiteral(valueSyntax);
+        }
+
+        /// <inheritdoc />
+        protected override StringValueNode ParseValue(string runtimeValue)
+        {
+            if (!ValidationRegex.IsMatch(runtimeValue))
+            {
+                throw ThrowHelper.HslType_ParseValue_IsInvalid(this);
+            }
+
+            return base.ParseValue(runtimeValue);
+        }
+
+        /// <inheritdoc />
+        public override bool TrySerialize(object? runtimeValue, out object? resultValue)
+        {
+            if (runtimeValue is null)
+            {
+                resultValue = null;
+                return true;
+            }
+
+            if (runtimeValue is string s &&
+                ValidationRegex.IsMatch(s))
+            {
+                resultValue = s;
+                return true;
+            }
+
+            resultValue = null;
+            return false;
+        }
+
+        /// <inheritdoc />
+        public override bool TryDeserialize(object? resultValue, out object? runtimeValue)
+        {
+            if (resultValue is null)
+            {
+                runtimeValue = null;
+                return true;
+            }
+
+            if (resultValue is string s &&
+                ValidationRegex.IsMatch(s))
+            {
+                runtimeValue = s;
+                return true;
+            }
+
+            runtimeValue = null;
+            return false;
+        }
+    }
+}