+++ src/StrawberryShake/Client/src/Core/Serialization/FloatSerializer.cs
@@ -2,7 +2,7 @@ namespace StrawberryShake.Serialization
 {
     public class FloatSerializer : ScalarSerializer<double>
     {
-        public FloatSerializer(string typeName = BuiltInTypeNames.Float)
+        public FloatSerializer(string typeName = BuiltInScalarNames.Float)
             : base(typeName)
         {
         }