+++ src/StrawberryShake/Client/src/Core/Serialization/IntSerializer.cs
@@ -2,7 +2,7 @@ namespace StrawberryShake.Serialization
 {
     public class IntSerializer : ScalarSerializer<int>
     {
-        public IntSerializer(string typeName = BuiltInTypeNames.Int)
+        public IntSerializer(string typeName = BuiltInScalarNames.Int)
             : base(typeName)
         {
         }