+++ src/StrawberryShake/CodeGeneration/src/CodeGeneration.CSharp/CSharpGeneratorResult.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using HotChocolate;
 
@@ -5,11 +6,28 @@ namespace StrawberryShake.CodeGeneration.CSharp
 {
     public class CSharpGeneratorResult
     {
+        public CSharpGeneratorResult()
+            : this(Array.Empty<CSharpDocument>(), Array.Empty<IError>())
+        {
+        }
+
         public CSharpGeneratorResult(
-            IReadOnlyList<CSharpDocument> cSharpDocuments,
+            IReadOnlyList<IError> errors)
+            : this(Array.Empty<CSharpDocument>(), errors)
+        {
+        }
+
+        public CSharpGeneratorResult(
+            IReadOnlyList<CSharpDocument> csharpDocuments)
+            : this(csharpDocuments, Array.Empty<IError>())
+        {
+        }
+
+        public CSharpGeneratorResult(
+            IReadOnlyList<CSharpDocument> csharpDocuments,
             IReadOnlyList<IError> errors)
         {
-            CSharpDocuments = cSharpDocuments;
+            CSharpDocuments = csharpDocuments;
             Errors = errors;
         }