+++ src/StrawberryShake/CodeGeneration/src/CodeGeneration.CSharp/Extensions/StringExtensions.cs
@@ -1,3 +1,4 @@
+using System.Linq;
 using HotChocolate;
 
 namespace StrawberryShake.CodeGeneration.CSharp
@@ -19,9 +20,19 @@ namespace StrawberryShake.CodeGeneration.CSharp
             return str + "<" + string.Join(", ", generics) + ">";
         }
 
-        public static string MakeNullable(this string str)
+        public static string WithGeneric(this string str, params RuntimeTypeInfo[] generics)
         {
-            return str + "?";
+            return str + "<" + string.Join(", ", generics.Select(x => x.ToString())) + ">";
+        }
+
+        public static string MakeNullable(this string str, bool isNullable = true)
+        {
+            if (isNullable)
+            {
+                return str + "?";
+            }
+
+            return str;
         }
     }
 }