+++ src/StrawberryShake/CodeGeneration/src/CodeGeneration.CSharp/Generators/InputValueFormatterGenerator.cs
@@ -1,92 +1,83 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using HotChocolate;
 using StrawberryShake.CodeGeneration.CSharp.Builders;
 using StrawberryShake.CodeGeneration.CSharp.Extensions;
 using StrawberryShake.CodeGeneration.Extensions;
 using StrawberryShake.Serialization;
+using IInputValueFormatter = StrawberryShake.Serialization.IInputValueFormatter;
 using static StrawberryShake.CodeGeneration.NamingConventions;
 using static StrawberryShake.CodeGeneration.Utilities.NameUtils;
-using IInputValueFormatter = StrawberryShake.Serialization.IInputValueFormatter;
 
 namespace StrawberryShake.CodeGeneration.CSharp
 {
-    public class InputValueFormatterGenerator : CodeGenerator<NamedTypeDescriptor>
+    public class InputValueFormatterGenerator : CodeGenerator<InputObjectTypeDescriptor>
     {
         private static readonly string _keyValuePair =
-            TypeNames.KeyValuePair.WithGeneric(
-                TypeNames.String,
-                TypeNames.Object.MakeNullable());
+            TypeNames.KeyValuePair.WithGeneric(TypeNames.String, TypeNames.Object.MakeNullable());
 
-        protected override bool CanHandle(NamedTypeDescriptor descriptor)
+        protected override bool CanHandle(InputObjectTypeDescriptor descriptor)
         {
-            return descriptor.Kind == TypeKind.InputType;
+            return true;
         }
 
         protected override void Generate(
             CodeWriter writer,
-            NamedTypeDescriptor namedTypeDescriptor,
+            InputObjectTypeDescriptor namedTypeDescriptor,
             out string fileName)
         {
-            fileName = InputValueFormatterFromType(namedTypeDescriptor);
+            const string serializerResolver = nameof(serializerResolver);
+            const string runtimeValue = nameof(runtimeValue);
+            const string value = nameof(value);
+
+            fileName = CreateInputValueFormatter(namedTypeDescriptor);
 
             NameString typeName = namedTypeDescriptor.Name;
-            ClassBuilder classBuilder = ClassBuilder.New()
+
+            ClassBuilder classBuilder = ClassBuilder
+                .New()
                 .SetName(fileName)
                 .AddImplements(TypeNames.IInputValueFormatter);
 
-            var neededSerializers = namedTypeDescriptor.Properties
-                .ToLookup(x => x.Type.Name)
-                .Select(x => x.First())
-                .ToDictionary(x => x.Type.Name);
+            var neededSerializers = namedTypeDescriptor
+                .Properties
+                .GroupBy(x => x.Type.Name)
+                .ToDictionary(x => x, x => x.First());
 
             //  Initialize Method
 
             CodeBlockBuilder initialize = classBuilder
                 .AddMethod("Initialize")
                 .SetPublic()
-                .AddParameter("serializerResolver", x => x.SetType(TypeNames.ISerializerResolver))
+                .AddParameter(serializerResolver, x => x.SetType(TypeNames.ISerializerResolver))
                 .AddBody();
 
             foreach (var property in neededSerializers.Values)
             {
-                if (property.Type.GetGraphQlTypeName()?.Value is { } name)
+                if (property.Type.GetName().Value is { } name)
                 {
-                    MethodCallBuilder call = MethodCallBuilder.New()
-                        .SetMethodName("serializerResolver." +
+                    var propertyName = GetFieldName(name) + "Formatter";
+
+                    initialize
+                        .AddAssigment(propertyName)
+                        .AddMethodCall()
+                        .SetMethodName(
+                            serializerResolver,
                             nameof(ISerializerResolver.GetInputValueFormatter))
-                        .AddArgument(name.AsStringToken() ?? "")
-                        .SetPrefix($"{GetFieldName(name)}Formatter = ");
+                        .AddArgument(name.AsStringToken());
 
-                    initialize.AddCode(call);
-                }
-                else
-                {
-                    throw new InvalidOperationException(
-                        $"Serializer for property {namedTypeDescriptor.Name}.{property.Name} " +
-                        $"could not be created. GraphQLTypeName was empty");
-                }
-            }
-
-            // Serializer Fields
-
-            foreach (var property in neededSerializers.Values)
-            {
-                if (property.Type.GetGraphQlTypeName()?.Value is { } name)
-                {
-                    FieldBuilder field = FieldBuilder.New()
-                        .SetName(GetFieldName(name) + "Formatter")
+                    classBuilder
+                        .AddField(propertyName)
                         .SetAccessModifier(AccessModifier.Private)
                         .SetType(TypeNames.IInputValueFormatter)
                         .SetValue("default!");
-
-                    classBuilder.AddField(field);
                 }
                 else
                 {
                     throw new InvalidOperationException(
                         $"Serializer for property {namedTypeDescriptor.Name}.{property.Name} " +
-                        $"could not be created. GraphQLTypeName was empty");
+                        "could not be created. GraphQLTypeName was empty");
                 }
             }
 
@@ -99,46 +90,44 @@ namespace StrawberryShake.CodeGeneration.CSharp
 
             // Format Method
 
-            classBuilder
+            ArrayBuilder arrayBuilder = classBuilder
                 .AddMethod(nameof(IInputValueFormatter.Format))
                 .SetPublic()
                 .SetReturnType(TypeNames.Object.MakeNullable())
-                .AddParameter(
-                    "runtimeValue",
-                    x => x.SetType(TypeNames.Object.MakeNullable()))
+                .AddParameter(runtimeValue, x => x.SetType(TypeNames.Object.MakeNullable()))
                 .AddBody()
-                .ArgumentException("runtimeValue", $"!(runtimeValue is {typeName} d)")
+                .ArgumentException(runtimeValue, $"!({runtimeValue} is {typeName} d)")
                 .AddEmptyLine()
                 .AddArray()
-                .SetPrefix("return ")
-                .ForEach(namedTypeDescriptor.Properties,
-                    (builder, property) =>
-                        builder
-                            .SetType(_keyValuePair)
-                            .AddAssigment(MethodCallBuilder.New()
-                                .SetPrefix("new ")
-                                .SetDetermineStatement(false)
-                                .SetMethodName(_keyValuePair)
-                                .AddArgument(property.Name.WithLowerFirstChar().AsStringToken())
-                                .AddArgument(MethodCallBuilder.New()
-                                    .SetMethodName($"Format{property.Name}")
-                                    .SetDetermineStatement(false)
-                                    .AddArgument($"d.{property.Name}"))));
+                .SetReturn();
 
-            // Serializer Methods
+            foreach (var property in namedTypeDescriptor.Properties)
+            {
+                var serializerMethodName = $"Format{property.Name}";
 
-            classBuilder.ForEach(
-                namedTypeDescriptor.Properties,
-                (builder, property) =>
-                    builder.AddMethod("Format" + property.Name)
-                        .AddParameter("value", x => x.SetType(property.Type.ToBuilder()))
-                        .SetReturnType(TypeNames.Object.MakeNullable())
-                        .SetPrivate()
-                        .AddCode(GenerateSerializer(property.Type, "value")));
+                arrayBuilder
+                    .SetType(_keyValuePair)
+                    .AddAssigment(MethodCallBuilder
+                        .Inline()
+                        .SetNew()
+                        .SetMethodName(_keyValuePair)
+                        .AddArgument(GetParameterName(property.Name).AsStringToken())
+                        .AddArgument(MethodCallBuilder
+                            .Inline()
+                            .SetMethodName(serializerMethodName)
+                            .AddArgument($"d.{property.Name}")));
+
+                classBuilder
+                    .AddMethod(serializerMethodName)
+                    .AddParameter(value, x => x.SetType(property.Type.ToBuilder()))
+                    .SetReturnType(TypeNames.Object.MakeNullable())
+                    .SetPrivate()
+                    .AddCode(GenerateSerializer(property.Type, value));
+            }
 
             CodeFileBuilder
                 .New()
-                .SetNamespace(namedTypeDescriptor.Namespace)
+                .SetNamespace(namedTypeDescriptor.RuntimeType.NamespaceWithoutGlobal)
                 .AddType(classBuilder)
                 .Build(writer);
         }
@@ -150,37 +139,56 @@ namespace StrawberryShake.CodeGeneration.CSharp
         {
             switch (typeDescriptor)
             {
-                case NamedTypeDescriptor descriptor:
-                    var type = descriptor.GetGraphQlTypeName()?.Value + "Formatter";
-                    var serializerName = GetFieldName(type);
-                    var methodCall = $"{serializerName}.Format({variableName})";
+                case INamedTypeDescriptor descriptor:
+                    var serializerName = GetFieldName(descriptor.GetName().Value) + "Formatter";
+                    MethodCallBuilder methodCall = MethodCallBuilder
+                        .New()
+                        .SetMethodName(serializerName, nameof(IInputValueFormatter.Format))
+                        .AddArgument(variableName);
+
                     return assignment == "return"
-                        ? CodeLineBuilder.From("return " + methodCall + ";")
-                        : CodeLineBuilder.From($"{assignment}.Add({methodCall});");
+                        ? methodCall.SetReturn()
+                        : MethodCallBuilder
+                            .New()
+                            .SetMethodName(assignment, nameof(List<object>.Add))
+                            .AddArgument(methodCall.SetDetermineStatement(false));
+
                 case NonNullTypeDescriptor descriptor:
-                    return CodeBlockBuilder.New()
+                    return CodeBlockBuilder
+                        .New()
                         .AddIf(x =>
                             x.SetCondition($"{variableName} == default")
                                 .AddCode(
                                     assignment == "return"
                                         ? CodeLineBuilder.From("return null;")
-                                        : CodeBlockBuilder.New()
+                                        : CodeBlockBuilder
+                                            .New()
                                             .AddLine($"{assignment}.Add(null);")
                                             .AddLine("continue;")))
                         .AddEmptyLine()
-                        .AddCode(GenerateSerializer(descriptor.InnerType(),
-                            variableName,
-                            assignment));
-                case ListTypeDescriptor descriptor:
-                    return CodeBlockBuilder.New()
                         .AddCode(
-                            CodeLineBuilder.From(
-                                $"var {variableName}_list = new " +
-                                $"{TypeNames.List.WithGeneric(TypeNames.Object.MakeNullable())}" + 
-                                "();"))
+                            GenerateSerializer(
+                                descriptor.InnerType(),
+                                variableName,
+                                assignment));
+
+                case ListTypeDescriptor descriptor:
+                    return CodeBlockBuilder
+                        .New()
+                        .AddCode(
+                            AssignmentBuilder
+                                .New()
+                                .SetLefthandSide($"var {variableName}_list")
+                                .SetRighthandSide(
+                                    MethodCallBuilder
+                                        .Inline()
+                                        .SetNew()
+                                        .SetMethodName(TypeNames.List)
+                                        .AddGeneric(TypeNames.Object.MakeNullable())))
                         .AddEmptyLine()
                         .AddCode(
-                            ForEachBuilder.New()
+                            ForEachBuilder
+                                .New()
                                 .SetLoopHeader($"var {variableName}_elm in {variableName}")
                                 .AddCode(
                                     GenerateSerializer(
@@ -190,7 +198,8 @@ namespace StrawberryShake.CodeGeneration.CSharp
                         .AddCode(
                             assignment == "return"
                                 ? CodeLineBuilder.From($"return {variableName}_list;")
-                                : CodeLineBuilder.From($"{assignment}.Add({variableName}_list);"));
+                                : CodeLineBuilder.From(
+                                    $"{assignment}.Add({variableName}_list);"));
                 default:
                     throw new InvalidOperationException();
             }